pyproject.toml
[project]
name = "playing-cards"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.2.1",
]
